requirements.txt
flask==3.0.2
openai==1.55.3
langfuse==2.60.2
python-dotenv==1.0.1
httpx[http2]==0.27.2
h2==4.1.0 